Undercharged random loot
Randomly placed loot takes on the charges of the placeholder item listed in the creature or area in which it's placed. In a
handful of places, these charges need to be adjusted so that the item doesn't show up broken, especially for items like the
Bone Amulet which don't recharge.

https://www.gibberlings3.net/forums/topic/33176-bone-talisman/