Swapping the dupe EXTRES4 fix
Fixpack addresses the duplicate random loot (EXTRES4, which randomizes to Cloak of Invisibility/Ogien's Scale) in the
Barbarian Camp and the Wyrm's Tooth interior. Fixpack does this by replacing the dupe in the Barbarian Camp with EXTRES1 (Belt
of Bones/Helm of Shouting/Stomper's Boots). However, looking at the charges on the placeholders, Fixpack should be replacing
the duplicate at Wyrm's Tooth so that the items come properly charged.

https://www.gibberlings3.net/forums/topic/33176-bone-talisman/