Find Traps fix... woof
This is a goddamn messy hack. Find Traps worked by pulsing an op 150, then delay casting a subspell every round. The
problem is that the op 146 for the delayed casting forces the original caster to interrupt and go through casting
motions every round, rendering them more or less useless for the 30-round duration. So, we instead have Find Traps change
the caster's SPECIFICS to a special value for the duration; traps will detect this value and force a custom Find Traps to
reveal themselves. Unfortunately, this means every door/region trap has to use a unique script so that they can set unique
area variables since there's no other way to prevent an infinite loop of revelation. It's ugly internally, but works
beautifully in game.